// src/generated-types/store-category.ts
export interface Id {
  id: string;
}

/** Message representing a status response */
export interface StatusResponse {
  success: boolean;
  message: string;
}

/** Message representing a list of store categories */
export interface StoreCategoryList {
  data: StoreCategoryWithTranslation[];
}

/** Message representing a store category */
export interface StoreCategory {
  id: string;
  slug: string;
  isAvailable: boolean;
  sortOrder: number;
}

/** Message representing a store category with translation */
export interface StoreCategoryWithTranslation {
  id: string;
  slug: string;
  isAvailable: boolean;
  sortOrder: number;
  title: string;
  description: string;
}

/** Message representing a store category translation */
export interface StoreCategoryTranslation {
  id: string;
  title: string;
  description: string;
  language: Language1;
}

/** Message representing a store category with its translation */
export interface StoreCategoryWithTranslations {
  category: StoreCategory | null;
  translation: StoreCategoryTranslation[];
}